Rearranging the datalayer class, erazing the datahandler class. Adding Plotting class that plots geometry shapes and images. Documented the changes in the datalayer and the new class.
