+++ hera/doc/source/measurements/GIS/examples.rst
@@ -5,3 +5,4 @@ GIS examples
     :maxdepth: 2
 
     examples/ManagingData
+    examples/Plotting
